backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
feedparser==6.0.10
geopy==2.4.1
ollama==0.1.7
python-multipart==0.0.6
python-dateutil==2.8.2
httpx>=0.25.2,<0.26.0  # For PocketBase HTTP client (compatible with ollama)
pytz  # For timezone handling

